+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## THE CHANGELOG ##
 
+### 0.9 ###
+
+- use messages handlers to auto-synchronize after an option update
+- use messages handlers to auto-synchronize after a main folder rename from website
+- after installation, open options page
+- little option page re-design
+
 ### 0.8 ###
 
 - **Unit tests fix**: avoid useless pings with empty `TeamKey`
@@ -8,7 +15,6 @@
 - **Unit tests fix**: FF now use auto-generated indexes too
 - **Unit tests fix**: offline detection so in this case, resetting local version number
 - avoid double synchronization with a protection, usefull in case of update > invalid_version wich may combine with a ping
-- use messages handlers to auto-synchronize after an option update
 - now use `api` instead of `ajax` in URL
 
 ### 0.7 ###